+++ claripy/backends/backend_vsa.py
@@ -5,7 +5,9 @@ import operator
 from functools import reduce
 
 from claripy.ast.base import Base
+from claripy.backends.backend import Backend
 from claripy.balancer import Balancer
+from claripy.errors import BackendError
 from claripy.operations import backend_operations_vsa_compliant, expression_set_operations
 from claripy.vsa import (
     BoolResult,
@@ -17,8 +19,6 @@ from claripy.vsa import (
     TrueResult,
     ValueSet,
 )
-
-from . import Backend, BackendError
 
 l = logging.getLogger("claripy.backends.backend_vsa")
